refactor(Lottos): WinningAnalyzer의 책임 lottos에게 전격 위임
